+++ src/base/index.ts
@@ -2,4 +2,5 @@ export * from './function';
 export * from './list';
 export * from './number';
 export * from './object';
+export * from './regexp';
 export * from './string';