+++ src/component/index.ts
@@ -1,0 +1,5 @@
+export { addEmitsForComponent } from './add-emits';
+export { addNameForComponent } from './add-name';
+export { extractClass } from './extract-class';
+export { extractEvent } from './extract-event';
+export { extractProps } from './extract-props';