+++ src/router/index.ts
@@ -1,0 +1,1 @@
+export { findRouteName } from './find-route-name';